Improve alignment of batch and non-batch queries to aide users
Batch queries require a followup call to Vulns API endpoint to get more details. Due to differences, to avoid confusion, if one package is called, it uses the one endpoint, if many are called, then it uses the latter and then aligns the formatting so each is a list of vulnerabilities of all packages input.