Remove reference to versions in main osv_query helper function
Right now the OSV queries will return all vulnerabilities for a package, which may affect many versions. As such, the entire list will reference all those listed. To avoid hidden filtering based upon a version, only packages are accepted in this function.

To simply test if one version is impacted, osv_query_1 would be more suitable for this purpose. There may be plans in the future to make this behaviour more obvious in the helper function. However, the purpose of this function is to get a complete set of vulnerabilities for packages provided, regardless of version so a black list can be made!