Update higher-level functions to use preparsing of lower-levels for queries
With this change, we do not have to rely on the limited parsing done in a higher level function. The main parsing work is down once at a much lower level and carried up.To make this easier, the outside from osv_query is an augmented data.table.

Minor alignment to OSV naming for columns as well as a bug in a missing method for summary data extraction during parsing.