Add filtering option to versions in high-level helper for queries
Since all versions are returned for vulnerability, this can help reduce it automatically. Otherwise, the user needs to select the versions they want to keep upon returns. If no versions are found, its an empty data.frame; if a version is found, it returns all impacted.

The default behaviour is kept to align with OSV methodology. For most circumstances, it will be preferable to keep all content as this is more flexbile for use-case differences.