Function to quickly detect if a package is listed has having a vulnerability.
This is a high level function that operates on just name and ecosystem inputs. It also does not provide version specific information.

Additional queries can be made, or future functionalty extensions, to allow more specific queries by version. However, to be robust, this would have to include both the 'versions' array as well as the 'range' array, which currently is not parsed. Some ecosystems only provide the ranges at this time. To avoid missing some vulnerabilities, just the package and ecosystem is used at this time to generate the vector.