Correct filtering method for osv_query
Was not performing as expected if multiple packages existed for one vulnerability. This ensures that only the provided package within the ecosystem, and possibly versions provided (NA if otherwise) will be reduced for all vulnerabilities returned.

This ensures returned values are a bit more predictable to the end users who may not expect other packages to be 'added' in the results when specifying specific package names. However, the default behaviour is to return all results since this is the normal behaviour from the API calls.