Implement pagination with latest httr2 helpers
Had to create an interator specific to OSV and for the different end points; furthermore, realignment had to be done so the object when collapsed was not just a set of vulnerabilities but maintained its relationship to the input vector. Otherwise it would be easy to mix vulnerabilities between ecosystems without the user knowing in an obvious way.

page_tokens were removed from mid-tier functions to enforce the method into the internal workings. There should be no reason for users to deal with page tokens after this release. If they really need the tokens, they will still exist in the response field.

With pagination, had to realign to 'response', 'queries', and 'vulns' list headers so that later parsing could remain unchanged.